// packages/vsr/src/generated/ResourceInputDocument.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FieldDocumentsInput } from "./FieldDocumentsInput";
import type { IndexDocument } from "./IndexDocument";
import type { ListConfigDocument } from "./ListConfigDocument";
import type { ManyToManyDocument } from "./ManyToManyDocument";
import type { ResourceAccessDocument } from "./ResourceAccessDocument";
import type { ResourceActionDocument } from "./ResourceActionDocument";
import type { ResourceApiInputDocument } from "./ResourceApiInputDocument";
import type { RoleRequirementsDocument } from "./RoleRequirementsDocument";
import type { RowPoliciesDocument } from "./RowPoliciesDocument";

export type ResourceInputDocument = { name: string, table: string | null, api_name: string | null, id_field: string | null, access: ResourceAccessDocument, roles: RoleRequirementsDocument, policies: RowPoliciesDocument, list: ListConfigDocument, api: ResourceApiInputDocument | null, use: Array<string>, indexes: Array<IndexDocument>, many_to_many: Array<ManyToManyDocument>, actions: Array<ResourceActionDocument>, fields: FieldDocumentsInput, };